common-module and dependency
