+++ Task-service/src/test/java/com/auth/taskservice/TaskServiceApplicationTests.java
@@ -1,0 +1,16 @@
+package com.auth.taskservice;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest
+class TaskServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
