+++ User-service/src/test/java/com/todolist/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.todolist;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
