+++ common-module/src/main/java/com/todolist/commonmodule/DTO/UserDto.java
@@ -1,0 +1,73 @@
+package com.todolist.commonmodule.DTO;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+
+public class UserDto {
+    @NotNull
+    @NotEmpty
+    private String firstName;
+
+    @NotNull
+    @NotEmpty
+    private String lastName;
+@NotNull
+    @NotEmpty
+    private String role;
+
+    @NotNull
+    @NotEmpty
+    private String password;
+    private String matchingPassword;
+
+    @NotNull
+    @NotEmpty
+    private String email;
+    @NotNull
+    @NotEmpty
+    private String phone;
+
+    public String getRole(){
+        return role;
+    }
+    public void setRole(String role){
+        this.role = role;
+    }
+    public String getPhone() {
+        return phone;
+    }
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+    public String getFirstName() {
+        return firstName;
+    }
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+    public String getLastName() {
+        return lastName;
+    }
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+    public String getPassword() {
+        return password;
+    }
+    public void setPassword(String password) {
+        this.password = password;
+    }
+    public String getMatchingPassword() {
+        return matchingPassword;
+    }
+    public void setMatchingPassword(String matchingPassword) {
+        this.matchingPassword = matchingPassword;
+    }
+    public String getEmail() {
+        return email;
+    }
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+}
