+++ common-module/src/main/java/com/todolist/commonmodule/security/Security/JwtUtil.java
@@ -1,0 +1,50 @@
+package com.todolist.commonmodule.security.Security;
+
+import io.jsonwebtoken.Claims;
+import org.springframework.stereotype.Component;
+
+import java.util.Date;
+import java.util.function.Function;
+import io.jsonwebtoken.Jwts;
+
+@Component
+public class JwtUtil {
+
+    private final String secretKey = "abcdefghijklmnopqrstuvwxyztuvwxyz1234567890";
+    private final long expirationTime = 1000 * 60 * 60; 
+
+    // יצירת הטוקן
+    public String generateToken(String email) {
+        return Jwts.builder()
+                .setSubject(email)
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + expirationTime))
+                .signWith(io.jsonwebtoken.SignatureAlgorithm.HS256, secretKey)
+                .compact();
+    }
+
+    public String extractUsername(String token) {
+        return extractClaim(token, Claims::getSubject);
+    }
+
+    private Claims extractAllClaims(String token) {
+        return Jwts.parser()
+                .setSigningKey(secretKey)
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
+        final Claims claims = extractAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    public boolean validateToken(String token, org.springframework.security.core.userdetails.UserDetails userDetails) {
+        final String username = extractUsername(token);
+        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
+    }
+
+    private boolean isTokenExpired(String token) {
+        return extractClaim(token, Claims::getExpiration).before(new Date());
+    }
+}
