+++ common-module/src/main/java/com/todolist/commonmodule/security/exceptions/UserAlreadyExistException.java
@@ -1,0 +1,12 @@
+package com.todolist.commonmodule.security.exceptions;
+
+public class UserAlreadyExistException extends RuntimeException {
+
+    public UserAlreadyExistException(String message) {
+        super(message);
+    }
+
+    public UserAlreadyExistException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
