--- src/main/java/com/task/Validation/PasswordValidator.java
@@ -1,31 +1,0 @@
-package com.task.Validation;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-public class PasswordValidator
-        implements ConstraintValidator<ValidPassword, String> {
-
-    private Pattern pattern;
-    private Matcher matcher;
-    private static final String PASSWORD_PATTERN =
-            "^(?=.*[A-Z])(?=.*[a-z])(?=.*[@#$%^&+=!])(?=.*[0-9]).{8,}$";
-
-    @Override
-    public void initialize(ValidPassword constraintAnnotation) {
-    }
-
-    @Override
-    public boolean isValid(String password, ConstraintValidatorContext context) {
-        return (validatePassword(password));
-    }
-
-    private boolean validatePassword(String password) {
-        pattern = Pattern.compile(PASSWORD_PATTERN);
-        matcher = pattern.matcher(password);
-        return matcher.matches();
-    }
-}
