--- src/main/java/com/task/Validation/ValidEmail.java
@@ -1,24 +1,0 @@
-package com.task.Validation;
-
-import com.task.Validation.EmailValidator;
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target({TYPE, FIELD, ANNOTATION_TYPE})
-@Retention(RUNTIME)
-@Constraint(validatedBy = EmailValidator.class)
-@Documented
-public @interface ValidEmail {
-    String message() default "Invalid email";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-}
