--- src/main/java/com/task/Validation/ValidPassword.java
@@ -1,24 +1,0 @@
-package com.task.Validation;
-
-import com.task.Validation.PasswordValidator;
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-@Target({TYPE, FIELD, ANNOTATION_TYPE})
-@Retention(RUNTIME)
-@Constraint(validatedBy = PasswordValidator.class)
-@Documented
-public @interface ValidPassword {
-    String message() default "Invalid password";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-}
