--- src/main/java/com/task/controllers/LoginController.java
@@ -1,19 +1,0 @@
-package com.task.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-@Controller
-
-public class LoginController {
-
-    @GetMapping("/loginPage")
-    public String loginPage() {
-        return "login";
-    }
-
-    @GetMapping("/registerPage")
-    public String showRegisterPage() {
-        return "registration";
-    }
-}
