--- src/main/java/com/task/entities/Users.java
@@ -1,73 +1,0 @@
-package com.task.entities;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.persistence.*;
-
-import javax.management.relation.Role;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "Users")
-public class Users {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonProperty("id")
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String username;
-    @Column(unique = false, nullable = false)
-    private String password;
-    @Column(unique = true, nullable = false)
-    private String email;
-    @Column(unique = true, nullable = false)
-    private String phone;
-    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<ListObject> user_list = new ArrayList<>();
-    private boolean enabled;
-    @Column(unique = false, nullable = false)
-    private String role;
-
-    public Users() {
-
-    }
-    public void setId(Long id) {
-        this.id = id;
-    }
-    public void setRole(String role) {
-        this.role = role;
-    }
-    public String getRole() {
-        return role;
-    }
-    public void setUsername(String username) {
-        this.username = username;
-    }
-    public void setPassword(String password) {
-        this.password = password;
-    }
-    public void setEmail(String email) {
-        this.email = email;
-    }
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public Long getId() {
-        return id;
-    }
-    public String getUsername() {
-        return username;
-    }
-    public String getPassword() {
-        return password;
-    }
-    public String getEmail() {
-        return email;
-    }
-    public String getPhone() {
-        return phone;
-    }
-
-}
