--- src/main/java/com/task/exceptions/UserAlreadyExistException.java
@@ -1,12 +1,0 @@
-package com.task.exceptions;
-
-public class UserAlreadyExistException extends RuntimeException {
-
-    public UserAlreadyExistException(String message) {
-        super(message);
-    }
-
-    public UserAlreadyExistException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
