--- src/main/resources/application.properties
@@ -1,42 +1,0 @@
-# this file contains key-values of configuration of the website
-# most of them are used directly by spring, you can add yours as well
-
-
-# SQL DATABASE
-spring.datasource.url=jdbc:mysql://localhost:3306/todolistdb?serverTimezone=UTC
-# replace here the login and password of your SQL server
-spring.datasource.username=root
-spring.datasource.password=
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.database-platform=org.hibernate.dialect.MySQL57Dialect
-spring.jpa.generate-ddl=true
-spring.jpa.show-sql=true
-
-# to show the queries in console
-logging.level.org.hibernate.SQL=debug
-logging.level.org.hibernate.type.descriptor.sql=trace
-
-## USING IN MEMORY DATABASE (INSTEAD OF SQL ABOVE)
-#spring.datasource.url=jdbc:h2:mem:ex4
-#spring.datasource.driverClassName=org.h2.Driver
-#spring.datasource.username=sa
-#spring.datasource.password=password
-#spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-#spring.jpa.defer-datasource-initialization=true
-
-# IN CASE WE WANT TO STORE SESSION DATA IN THE DB
-# Note: this is not compatible with Session Listeners
-spring.session.store-type=jdbc
-spring.session.jdbc.initialize-schema=always
-# Sets the session timeout
-server.servlet.session.timeout=30m
-server.servlet.session.cookie.name=mySessionCookie
-
-spring.freemarker.suffix=.html
-spring.h2.console.enabled=true
-
-demo.coursename=Internet Course
-
-# define the path of the error page
-server.error.path=/error
